Moved definition of UserModel to bindings/UserModel.hpp + added template parameters. UserModel will also be used by the Python bindings
